DUGK-39 Refactored Menu.tsx, so Drawer and DrawerHeader is their own components (#33)
