+++ client/src/page/Menu.tsx
@@ -1,81 +1,22 @@
 import * as React from 'react';
-import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles';
+import { useTheme } from '@mui/material/styles';
 import Box from '@mui/material/Box';
-import MuiDrawer from '@mui/material/Drawer';
-import List from '@mui/material/List';
-import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import ChevronRightIcon from '@mui/icons-material/ChevronRight';
-import MenuToolbar, { transition } from './MenuToolbar';
-import MenuItems from './MenuItems';
+import MenuToolbar from './MenuToolbar';
+import DrawerComponent from './DrawerComponent';
 
 const drawerWidth = 240;
-
-const openedMixin = (theme: Theme): CSSObject => ({
-  width: drawerWidth,
-  transition: transition({ theme, open: true }),
-  overflowX: 'hidden',
-});
-
-const closedMixin = (theme: Theme): CSSObject => ({
-  transition: transition({ theme, open: false }),
-  overflowX: 'hidden',
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up('sm')]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-
-const DrawerHeader = styled('div')(({ theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'flex-end',
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-}));
-
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})(({ theme, open }) => ({
-  width: drawerWidth,
-  flexShrink: 0,
-  whiteSpace: 'nowrap',
-  boxSizing: 'border-box',
-  ...(open && {
-    ...openedMixin(theme),
-    '& .MuiDrawer-paper': openedMixin(theme),
-  }),
-  ...(!open && {
-    ...closedMixin(theme),
-    '& .MuiDrawer-paper': closedMixin(theme),
-  }),
-}));
 
 function MiniDrawer() {
   const theme = useTheme();
   const [open, setOpen] = React.useState<boolean>(false);
-  const [anchorElUser, setAnchorElUser] = React.useState<
-    (EventTarget & HTMLButtonElement) | null
-  >(null);
+  const [anchorElUser, setAnchorElUser] =
+    React.useState<HTMLButtonElement | null>(null);
 
-  const handleCloseUserMenu = () => {
-    setAnchorElUser(null);
-  };
-
-  // TODO: Fix the user, ones authentication is implemented
-  const handleOpenUserMenu = (event: React.MouseEvent<HTMLButtonElement>) => {
+  const handleCloseUserMenu = () => setAnchorElUser(null);
+  const handleOpenUserMenu = (event: React.MouseEvent<HTMLButtonElement>) =>
     setAnchorElUser(event.currentTarget);
-  };
-
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
-
-  const handleDrawerClose = () => {
-    setOpen(false);
-  };
+  const handleDrawerOpen = () => setOpen(true);
+  const handleDrawerClose = () => setOpen(false);
 
   return (
     <Box sx={{ display: 'flex' }}>
@@ -87,21 +28,11 @@ function MiniDrawer() {
         handleOpenUserMenu={handleOpenUserMenu}
         anchorElUser={anchorElUser}
       />
-      <Drawer variant="permanent" open={open}>
-        <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === 'rtl' ? (
-              <ChevronRightIcon />
-            ) : (
-              <ChevronLeftIcon />
-            )}
-          </IconButton>
-        </DrawerHeader>
-        <Divider />
-        <List>
-          <MenuItems />
-        </List>
-      </Drawer>
+      <DrawerComponent
+        open={open}
+        theme={theme}
+        handleDrawerClose={handleDrawerClose}
+      />
     </Box>
   );
 }
